requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
python-dotenv
openai
PyMuPDF
marshmallow
marshmallow-sqlalchemy
requests
